Logout integrated
