+++ BobaSpot/BobaSpot/frontend/src/contents/landing/Home.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from "./Navbar"
+import Navbar from "../../components/NavBar/NavBar"
 import About from "./About"
 import Map from "./Map"
 import BannerBgr from "./Assets/landing_banner.png";
